Move GPU based rendering logic into gpu.rs
This way the code is a bit more readeable.
